pyproject.toml
[tool.poetry]
name = "meeting-to-modules"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "mtm", from = "src"}]

[tool.poetry.scripts]
mtm = "mtm.cli:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "~0.12"}
pydantic = ">=2.0,<3.0"
pyyaml = "~6.0"
python-frontmatter = "~1.0"
markdown = "~3.5"
pdfminer-six = "^20221105"
python-docx = "~1.1"
scikit-learn = "^1.4"
nltk = "~3.9"
rapidfuzz = "~3.9"
jinja2 = "~3.1"
sqlite-utils = "~3.36"
toml = "~0.10"
rich = "~13.7"
python-dateutil = "~2.9"
unidecode = "~1.3"
python-slugify = "~8.0"
fastapi = "^0.115"
uvicorn = {extras = ["standard"], version = "^0.32"}

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-cov = "*"
ruff = "*"
mypy = "*"
types-python-dateutil = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

